Actually, use findMemberNoDealias
Contrary to what I mentioned in the last commit message, we can't
actually use lookupClassSymbol, because it causes us to miss a ton of
errors on the fast path. When the fast path rolls around, if we
previously reported a "redefining constant X" error, the
`lookupClassSymbol` would cause us to always look up the symbol kind
that made it not an error, depriving the fast path the chance to report
the error that was raised earlier.

In any case, since on the first path we do not allow class definitions
through aliases, we can simply use `findMemberNoDealias` all the time.

For example, this is not allowed, even on the slow path:

    module Outer
      module Namespace; end
      AliasToNamespace = Namespace
    end
    class Outer::AliasToNamespace::A; end

which means that we do not have to allow looking through aliases here in
namer.