Don't dealias in constantResolutionFailed
If we see an example like

    A = B
    A::X

We need to say that the `resolutionScopes` for `A::X` is `[A]`.
Previously, the `dealias` here caused us to say that the scope was
`[noSymbol()]` for `X`, which is wrong--we would have looked for `X` in
`A` if we had known what `A` was an alias to.

If we do know what `A` is an alias to, it's still fine to say that the
resolution scope was `A`--we can always dealias it downstream if need be.