Fix bug with order
Even though we populated the symbol table correctly, certain fast paths
would make the error message switch from show "as a type member" to "as
a static field" because the fast path edit would take a different branch
that had this bug (`oldSym` vs `existingTypeMember`).

(The other one was fine because we were hard-coding `FieldOrStaticField`
at the call site, so the order of `renamedSym` and `sym` didn't matter).