Put the constant kind in the redefinition error (#6263)
* Force class alias jobs to run

* Fix bug with scope lookup

Apart from the renames (I struggled more than I would like to admit
keeping this code straight while debugging, because it is highly
recursive), the main change is the `lookupClassSymbol`.

That call makes sure that it finds the right symbol kind in the presence
of previous mangle renames. Note that there are only two places where we
call `ensureScopeIsClass`, and the other one (in `insertStaticField`)
also uses `lookupStaticFieldSymbol` instead of `findMember`.

Without this change, what would happen is that given a test case like
this:

    module A::B::C
      def foo; end
    end

    A::B = A::D

On the incremental path, the `findMember` call would look up `A::B` as a
static field. And now that the static field properly stores the alias to
`StubModule`, class `C` would get entered on `StubModule` instead of on
the existing `A::B::C` class symbol.

(This test exists as `nested_module_then_nonexistent_name.rb`.)

* Actually, use findMemberNoDealias

Contrary to what I mentioned in the last commit message, we can't
actually use lookupClassSymbol, because it causes us to miss a ton of
errors on the fast path. When the fast path rolls around, if we
previously reported a "redefining constant X" error, the
`lookupClassSymbol` would cause us to always look up the symbol kind
that made it not an error, depriving the fast path the chance to report
the error that was raised earlier.

In any case, since on the first path we do not allow class definitions
through aliases, we can simply use `findMemberNoDealias` all the time.

For example, this is not allowed, even on the slow path:

    module Outer
      module Namespace; end
      AliasToNamespace = Namespace
    end
    class Outer::AliasToNamespace::A; end

which means that we do not have to allow looking through aliases here in
namer.

* Don't dealias in constantResolutionFailed

If we see an example like

    A = B
    A::X

We need to say that the `resolutionScopes` for `A::X` is `[A]`.
Previously, the `dealias` here caused us to say that the scope was
`[noSymbol()]` for `X`, which is wrong--we would have looked for `X` in
`A` if we had known what `A` was an alias to.

If we do know what `A` is an alias to, it's still fine to say that the
resolution scope was `A`--we can always dealias it downstream if need be.

* Add test showing master behavior

* Show new behavior

* Add test showing master

* Capture branch behavior

* Don't show StubModule in error messages

* whoops forgot error assertion

* exp files

* pre-work: errorLoc Loc -> LocOffsets

* pre-work: Always take prev as a SymbolRef

* Put the redefinition kind in the error message

* Update this 4012 reference

* Make every existing error message more specific

* Pass around NameRef, not strings

* Fix bug with order

Even though we populated the symbol table correctly, certain fast paths
would make the error message switch from show "as a type member" to "as
a static field" because the fast path edit would take a different branch
that had this bug (`oldSym` vs `existingTypeMember`).

(The other one was fine because we were hard-coding `FieldOrStaticField`
at the call site, so the order of `renamedSym` and `sym` didn't matter).

* update cli

* Fix this test