Update test expectations
Most of these are just wording changes of existing error messages, but a
handful of tests (like `nested_static_field.rb`) actually change
meaning. I think that the change is clear (maybe more clear than before)
and only present in files that had errors previously, so it's fine.

While in `nested_static_field.rb` the error feels out of order,
overwhelmingly I think this error is going to be hit when two separate
files combine to have the buggy code, so the confusion of
seemingly-out-of-order execution will likely not be that common in
practice.