Allow `requires_ancestor` block to return a `T.class_of` object (#6299)
* Add test case calling `requires_ancestor` with `T.class_of`

* Set return type of requires_ancestor proc to void

Historically, the `requires_ancestor` proc argument has expected a
`Module` as the return type. However, if `T.class_of(...)` does not
satisfy this return type. Instead, we mark the return type as `void` and
implement code-level checks in another commit.

* Move `isT` method to Helpers`

* Pass singleton class to `recordRequiredAncestor`

Refactor the `resolveRequiredAncestorsJob` method to always pass the
required ancestor, whether a ConstantLit or a singleton class, as a
`ClassOrModuleRef`.

* Allow "generic" ancestor if it is a singleton class

* Update requires_ancestor documentation

* Move helper methods to anonymous namespace on resolver.cc

* Remove unnecessary nullptr check

* Remove unnecessary copy and variable instantiation

* Remove another unnecessary nullptr check

* Linting

* Add new tests