Update a snapshot test
The behavior in the class_and_alias.rb file is showing what happens when
an assignment like `X = 1` comes before or after a definition like
`class Y; end`.

Before, the behavior was that the first definition was always mangled.
Now the behavior is that the class or module definition is always
mangled.