Smash all `vector<SymbolHash>`'s into one big vector
This is a net-negative diff on its own, and also lets us avoid writing
*more* code in these places when we move static fields and type members
onto the fast path.

Whiel most of the changes are purely mechanical, the changes in
LSPFileUpdates.cc are slightly less obviously correct. We used to have
different set difference logic based on whether the boolean
`lspExperimentalFastPathEnabled` was enabled and whether the symbol kind
was static field or not.

But in the case where `validateIdenticalFingerprints` held (e.g., for
static fields), `c_set_difference` and `c_set_symmetric_difference`
should behave identically, especially considering that we de-dup the
hashes after the difference.

So while there's a slight difference in behavior, effectively there's no
change.

**Note** that this is only the `localSymbolTableHashes`. The vectors in
FoundDefHashes are still split out as separate vectors, because there is
not a single `FoundDefHash` kind (instead, there are `FoundMethodHash`
and `FoundFieldHash` structures, which store diffent information). Also,
the process of deleting methods given their `FoundDefinition` hash
requires knowing the kind of symbol we're operating on. So those vectors
will likely stay as is for the time being.