Symbolize type_member assignments (and other fixes) (#6411)
* pre-work: Just take one argument

The others we can derive from the one.

* pre-work: Move this snippet outside the `if`

It doesn't actually depend on any of the `type_member`'s args.

It was written where it was because also it wasn't previously required
except inside that block, but that will change in the next commit.

* Symbolize type_member assignments

We were previously not converting `UnresolvedConstantLit` nodes to
`ConstantLit` nodes in the LHS of a `type_member` assignment.

This was being masked by `builder_walk.cc` calling
`unresolvedIdent2Local` to effectively do the constant resolution lazily
in CFG builder. However, that constant resolution didn't take into
account mangle renames, which meant that sometimes `X = type_member`
would resolve to a `class ::Wrapper::X` symbol if the `type_member`
symbol had been redefined later in the file.

This incorrect resolution caused an ENFORCE failure in
`builder_entry.cc` that asserted that the `aliases` list only contains
static fields and type members (in the buggy version, it also could
contain class-or-module symbols).

* Add a test

* Remove an error

This error is spurious. It only used to appear because of how we decided
to recover from the namer redefinition error.

* Remove ENFORCE, because it wasn't true

* Make findSymbols and symbolizeTrees symmetric

These two `postTransformAssign` methods are nearly identical (the only
difference is that there is some extra code to handle instance variables
in the `findSymbols` case--namer doesn't need to deal with instance
variables in `symbolizeTrees`, so it's absent there).

This is pre-work for the next commit.

* pre-work: Add extra handling for weird scopes with `type_member`

* Add snapshot showing master behavior

* Show new behavior

* Add another test (top-level crasher)

* Capture bad behavior in a test, and record an issue for the future