Smash all `vector<SymbolHash>`'s into one big vector (#6421)
<!-- (optional) Explain your change, focusing on the details of the solution. This is a great place to call out user-visible changes. -->


### Motivation
<!-- Why make this change? Describe the problem, not the solution. This can also be a link to an issue. -->

This is a net-negative diff on its own, and also lets us avoid writing
*more* code in these places when we move static fields and type members
onto the fast path.

While most of the changes are purely mechanical, the changes in
`LSPFileUpdates.cc` are slightly less obviously correct. We used to have
different set difference logic based on whether the boolean
`lspExperimentalFastPathEnabled` was enabled and whether the symbol kind
was static field or not.

But in the case where `validateIdenticalFingerprints` held (e.g., for
static fields), `c_set_difference` and `c_set_symmetric_difference` should
behave identically, especially considering that we de-dup the hashes after
the difference.

So while there's a slight difference in behavior, effectively there's no change.

**Note** that this is only the `localSymbolTableHashes`. The vectors in
`FoundDefHashes` are still split out as separate vectors, because there is
not a single `FoundDefHash` kind (instead, there are `FoundMethodHash` and
`FoundFieldHash` structures, which store different information). Also, the
process of deleting methods given their `FoundDefinition` hash requires
knowing the kind of symbol we're operating on. So those vectors will likely
stay as is for the time being.


### Test plan
<!-- If you did not write tests for this change, replace the message below explaining why not. Why we should be confident this change is correct? If you changed the website, please include a screenshot of the proposed changes. -->

See included automated tests.