pre-work: Add "deleteSymbolViaFullNameHash" (kind-agnostic)
The diff is ugly, but this is basically just factoring out the common
parts of `delete{Field,Method}ViaFullNameHash` to a helper function,
called `deleteSymbolViaFullNameHash`. The only part remaining that is
not shared by those two `delete*` methods is the logic to compute the
`owner` symbol whose members should be looped.

This works by the observation that a `NameRef` already uniquely
identifies a Symbol in the `members` list, so comparing name hashes for
equality is good enough.

Note that we unwrap certain kinds of unique names, but that's
intentional, as it simply means that we'll delete more symbols (e.g.,
those that were created due to a mangle rename). But names like `@foo`
and `foo` and `Foo` should has to different values, and if they don't,
the only problem is for the symbol kinds that we can't re-enter, which
is just class or module symbols, so we can skip over those and treat the
ones we don't skip over agnostically with `SymbolRef`.