update all symbol-table exp tests
Note that the only thing that's changing is the locs

This is because not only does this change affect the fast path, but it
also uses the class `declLoc` for the loc of the symbol, instead of the
location of the constant literal in the class def. This makes more error
messages consistent (e.g., if there is an abstract instance or
singleton class method, the error location should be the same for both).