pre-work: name -> nameForErrors
The thing we're ensuring is a class is the scope only, not the name of
the constant itself. For example:

    class A::B::C
      X::Y::Z = 1
    end

This leads to two calls to this method:

    ensureScopeIsClass(A::B, "C")
    ensureScopeIsClass(X::Y, "Z")

Only `A::B` and `X::Y` are checked for being a class. In this case,
`X::Y::Z` is not a class, and that's OK—the name passed to this method
is only for error messages, so that we can say more about the cause of
the error.