Run packager PropagateVisibility pass on fast path
We need to include `__package.rb` files on all fast path edits, so that
when we remove and re-add static fields and type members, the
PropagateVisibility has a chance to set the `isExported` flag on the new
definitions.

At this point, the existing `VisibilityChecker::run` algorithm already
works correctly in incremental mode (assuming that the `__package.rb`
file did not actually change, which is guaranteed by canTakeFastPath).

In the future, when it's possible to take the fast path for
`__package.rb` changes, we'll ahve to update `VisibilityChecker::run` to
handle incremental edits, but we would have already had to do that, even
if there were separate `VisibilityChecker::{run,runIncremental}`
methods.