Require block arg type in signature
This commit ensures explicit block arguments must be defined in the
signature:

    sig {void}
    def foo(&blk); end
             ^^^ error: Malformed `sig`. Type not specified for argument `blk`

The following will not produce an error:

    sig {params(blk: T.proc.void).void}
    def foo(&blk); end

Every method has a "synthetic" block argument if an explicit one isn't
defined, for which an argument type **shouldn't** be required in the
signature. We were previously determining if a block arg was synthetic
or not based on its name reference. However, block arguments always have
the name reference `<blk>` whether they are explicitly defined or
synthetically added. This might not be intended, but is currently how it
works.

This commit swaps the name comparison for `isSyntheticBlockArgument`,
which has a more robust way of determining if the block is synthetic or
not, based on whether a `Loc` exists. Synthetic blocks use
`Loc::none()`.