Require block arg type in signature (#6523)
* Remove explicit &blk from T::Props.const

In a later commit we'll enforce argument types for explicit blocks.
Given this block argument isn't used by the `T::Props.const` method,
let's remove it rather than declaring an argument type.

* Remove explicit &blk from infer/generics/options

In a later commit we'll enforce argument types for explicit blocks.
Given these block arguments aren't used by the methods, let's remove
them rather than declaring argument types.

* Add block argument definitions to openssl

In a later commit we'll enforce argument types for explicit blocks.
This commit updates the openssl RBI to declare argument types for
methods that have explicit blocks.

* Add tests for arg types not specified

* Require block arg type in signature

This commit ensures explicit block arguments must be defined in the
signature:

    sig {void}
    def foo(&blk); end
             ^^^ error: Malformed `sig`. Type not specified for argument `blk`

The following will not produce an error:

    sig {params(blk: T.proc.void).void}
    def foo(&blk); end

Every method has a "synthetic" block argument if an explicit one isn't
defined, for which an argument type **shouldn't** be required in the
signature. We were previously determining if a block arg was synthetic
or not based on its name reference. However, block arguments always have
the name reference `<blk>` whether they are explicitly defined or
synthetically added. This might not be intended, but is currently how it
works.

This commit swaps the name comparison for `isSyntheticBlockArgument`,
which has a more robust way of determining if the block is synthetic or
not, based on whether a `Loc` exists. Synthetic blocks use
`Loc::none()`.

* squash! Remove explicit &blk from T::Props.const

These block arguments also aren't used by the methods.

* fixup! Require block arg type in signature

* fixup! Require block arg type in signature