Enter proper FoundClass definitions for all scopes (#6521)
* Enter proper FoundClass definitions for all scopes

This lets us

- Remove FoundClassRef
- Delete a bunch of code in namer to handle "what if the scope of a
  class or static field is actually a static field"
- Ensure that class symbols are _never_ mangled, properly this time.

* Skip `addLoc` for `FoundClass` with kind `Unknown`

* Add ENFORCE to enterClassSymbol

* Rename klassSymbol -> symbol

Previously, we had two variables because one was a SymbolRef and one was
a ClassOrModuleRef. Now this function works entirely on ClassOrModuleRef
and so doesn't need to have two separate variable names.

* "Can't nest under" -> "Cannot init by assign" in tests

* Use singletonClass declLoc for AttachedClass

These AttachedClass locs previously used the constant literal's
location. These locs shouldn't matter in practice.

* Some classes don't get static-init methods anymore

These classes didn't have class def bodies to begin with, so this should
be fine.

* One weird static field scope test

This test has code where `B::C = 1` comes before `module B`.

This now causes the `<AttachedClass>` and `<static-init>` symbols to get
entered before the `module B` has a chance to overwrite them with the
loc of the `ClassDef`.

We may want to change this in the future, but I'm `¯\_(ツ)_/¯` for now.