Add fast path test that currently fails
I first noticed this bug while working on #6422.

This test currently shows incorrect behavior. As far as I can tell, this
has been incorrect for as long as Sorbet has had a fast path mode. The
fact that this bug is only appearing now means that maybe we can
tolerate it a while longer?

In any case, we will want to fix this eventually.