Remove TODO(jez) by adding test case
If we attempt to delete the mentioned block of code (like the comment
suggested might be possible), this is what the symbol table looks like
when we re-run the fast path but don't run the incremental namer:

    contents before Namer::run
        method <C <U PrivateOverloads>>#<O <U foo> $1> : private (x, <blk>) -> Integer @ Loc {file=/home/jez/.dotfiles/vim/bundle/vim-sorbet/.empty/foo.rb start=7:3 end=7:44}
        method <C <U PrivateOverloads>>#<U foo> : private (<blk>) -> NilClass @ Loc {file=/home/jez/.dotfiles/vim/bundle/vim-sorbet/.empty/foo.rb start=8:11 end=8:25}
        method <C <U PrivateOverloads>>#<V <U foo> $1> : private (x, <blk>) @ Loc {file=/home/jez/.dotfiles/vim/bundle/vim-sorbet/.empty/foo.rb start=8:11 end=8:25}

    contents after Namer::run, before Resolver::runIncremental
        method <C <U PrivateOverloads>>#<O <U foo> $1> : private (x, <blk>) -> Integer @ Loc {file=/home/jez/.dotfiles/vim/bundle/vim-sorbet/.empty/foo.rb start=7:3 end=7:44}
        method <C <U PrivateOverloads>>#<U foo> : private (<blk>) -> NilClass @ Loc {file=/home/jez/.dotfiles/vim/bundle/vim-sorbet/.empty/foo.rb start=8:11 end=8:25}
        method <C <U PrivateOverloads>>#<V <U foo> $1> (x, <blk>) @ Loc {file=/home/jez/.dotfiles/vim/bundle/vim-sorbet/.empty/foo.rb start=8:11 end=8:25}

    contents after Resolver::runIncremental
        method <C <U PrivateOverloads>>#<O <U foo> $1> (x, <blk>) -> Integer @ Loc {file=/home/jez/.dotfiles/vim/bundle/vim-sorbet/.empty/foo.rb start=7:3 end=7:44}
        method <C <U PrivateOverloads>>#<U foo> (<blk>) -> NilClass @ Loc {file=/home/jez/.dotfiles/vim/bundle/vim-sorbet/.empty/foo.rb start=8:11 end=8:25}
        method <C <U PrivateOverloads>>#<V <U foo> $1> (x, <blk>) @ Loc {file=/home/jez/.dotfiles/vim/bundle/vim-sorbet/.empty/foo.rb start=8:11 end=8:25}

What that means is that simply answering a hover or definition request
would mutate the visibility of a method, absent any actual file changes.

In fact, with just the tests that exist as of #6422, we could have
deleted the mentioned block of code and the suite would have passed, but
the lingering bug would have been added.

I've verified that when based on the changes in #6422 AND after deleting
that block of code, the test case included in this PR fails. So this
test should prevent someone from unknowingly deleting that block in the
future (or guide them if they attempt to implement the suggestion in the
comment).

(I don't happen to think that doing that refactor is particularly urgent
right now, because the hack works ~well enough.)