Factor out errorFlusher argument to runSlowPath
The `runFastPath` and some of the methods it calls take an explicit
`errorFlusher` argument, but `runSlowPath` doesn't. In a future change,
we're going to want to start calling those same APIs from `runSlowPath`.
So it's convenient to make these two methods have basically the same
signature.

This should be a no-op, because every instance of an errorFlusher used
in this file is always constructed in the same way with values that
would be the same at runtime to those that are used in this PR.