Type member and static field edits take fast path (#6422)
* nonDeletableDefinitions -> deletableDefinitions

Only classes are non-deletable now.

* Add FoundTypeMemberHash

* Add FoundStaticFieldHash

* Static fields and type members on fast path

* Change the logic for when we sometimes don't run incremental namer

The comment in the implementation below describes more fully.

You can read more about the motivation for this change in the commit for the PR that added it:

<https://github.com/sorbet/sorbet/pull/5989/commits/78b8f6f9b2b659b0da7a0c6f15cc7f20412e0142>

Simply deleting the check added in that commit still didn't work,
because the `CanPreemptSlowPathWithCodeAction` test still failed.
(It seems the changes we made in #6176 to skip re-running infer if it
has already been typechecked on a newer epoch did not apply, because the
epoch for a retypecheck command is always manually set to `0`, so it
looked like the last type check was older, not newer than the slow path
that might have been pre-empted. I think that epoch of `0` mostly makes
sense because we wouldn't want any of the errors that a `retypecheck`
produces to ever get reported. Only the errors in response to real edits
should get reported. And we use epochs to decide which is the latest
version of a file to report errors for.)

But as for the problem alluded to in the original commit, the other half
of the change was fixed; fast path edits can once again pre-empt slow
path edits (that is #6176), which means that the `.empty()` check was
actually over eager, and we can be okay handling all fast path edits
using incremental namer now.

* Run finalizeSymbols on fast path

This finalizes the type members we've created on the incremental namer.

* Make finalizeSymbols idempotent on fast path

We don't yet run finalizeSymbols on the fast path, but when we do, we
need to make sure that we don't call `addMixin` over-eagerly when
handling `mixes_in_class_methods`.

* Delete all type members

* Also delete type_templates

* Special handling for the type_template class aliases

* Also need to delete the static field forwarder

* Exclude resultType from ClassOrModule::hash

* Reset some resultType's

Manually setting `resultType` for the `ClassOrModule` symbols to
`nullptr` and to a `LambdaParam(tp, todo, todo)` for the `AttachedClass`
type template mimics what would be the state had we just called
`insertClass` for the first time, and removes all references of the
newly-deleted type members from the containing class.

This forces resolver to recompute these types, taking into account any
deleted and added type members.

* Run packager PropagateVisibility pass on fast path

We need to include `__package.rb` files on all fast path edits, so that
when we remove and re-add static fields and type members, the
PropagateVisibility has a chance to set the `isExported` flag on the new
definitions.

At this point, the existing `VisibilityChecker::run` algorithm already
works correctly in incremental mode (assuming that the `__package.rb`
file did not actually change, which is guaranteed by canTakeFastPath).

In the future, when it's possible to take the fast path for
`__package.rb` changes, we'll ahve to update `VisibilityChecker::run` to
handle incremental edits, but we would have already had to do that, even
if there were separate `VisibilityChecker::{run,runIncremental}`
methods.

* Detect loop in PropagateVisibility in different way

* assert-slow-path -> assert-fast-path

* Remove disable-fast-path from type_member tests 🎉

* New tests

* Add classOrModuleShapeHash

See context in the ClassOrModule::hash comment below.

This makes the `type_member_redeclare_parent` test also typecheck
`__2.rb` on the fast path (there's still a bug, because that test
currently should retypecheck `__3.rb` but it doesn't, in a way that's
hard to fix).

* `deletable*` -> `retypecheckable*`

After using this map in the previous commit to track ClassOrModule
symbols too (for certain kinds of type member changes), "deletable"
didn't make sense anymore. Luckily nowhere it was used ever actually
relied on the symbols in there being deletable. It was only used to
figure out which files to retypecheck.

If you have another name please lmk.

* Actually, it's easy to be correct by construction

* Suggestions from code review

* remove_const_from_generic -> remove_static_field_from_generic_class

For consistency with the add_static_field_to_generic_class test.