Fix undefined behavior in signature override validation
To validate signatures overrides, for each required and optional parameter
from the super method we check if the corresponding parameter in the
overriding method is defined with a valid type.

Before this change, the override validation was checking if the corresponding
parameter was both in the `required` and `optional` list.
But in the case where the parameter was in the `required` list, we shouldn't
look into the `optional` one. If we do, we're going to compare the wrong iterator
against the wrong `end` and get an undefined behavior which could in some case
return true.

Reading the code it seems that what we actually want to do is check the type
if the parameter from the super method is found in the overriding method.

Signed-off-by: Alexandre Terrasa <alexandre.terrasa@shopify.com>