deletableDefinitions -> nonClassConstants (#6603)
This name doesn't make much sense anymore--it doesn't have methods nor
fields, which are also deletable definitions. Maybe its usefulness will
one day go away if we make edits to classes take the fast path too. But
for the time being, it's here to stay so we may as well make the name
change.