Include thread_mattr_accessor in Mattr dsl
[`thread_mattr_accessor`][thread_mattr_accessor] (and it's alias
`thread_cattr_accessor`) defines the same set of methods as
`mattr_accessor` and `cattr_accessor` (support for those was added in
https://github.com/sorbet/sorbet/pull/1252). I was recently confused
about why one of them worked and the other errored.

[thread_mattr_accessor]: https://api.rubyonrails.org/classes/Module.html#method-i-thread_mattr_accessor

This commit adds support for these additional methods. They might not be
used as often as `mattr_accessor` and friends, but I thought it might
still be worth adding since it's such a small change.