hack pipeline runner to mimic running under autogen
The previous commit had one problem: our `autogen.exp` files are
generated with `GlobalState::runningUnderAutogen` set to true thanks
to:

https://github.com/sorbet/sorbet/blob/50dcfd910ca8db9fa10c98b0edcad8868c19632c/main/realmain.cc#L497-L499

Our test harness, then, needs to set the same variable.  We could
theoretically set it for the duration of the test harness, but as it
mainly affects the rewriter pass, we might as well just set it for the
duration of the rewriter pass.