Remove this test
This test is a test of whether we report multiple def problems on the
slow path, which is covered by our existing testdata/cli tests. The
specific behavior relating to incrementality is covered by the other two
tests.

(As a thermometer to see why this test isn't all that useful: if we made
adding modules take the fast path, we'd have to keep editing this test.
So this is more a test that adding a module takes the slow path than it
is a test of multiple behavior defs.)