Improve Sorbet runtime caller frame matching logic (#6664)
Currently Sorbet runtime matches caller locations against what it
assumes are Sorbet runtime source file locations by using a regular
expression. The regular expression expects the Sorbet runtime `lib`
folder to be directly placed in a folder whose name includes the string
`"sorbet-runtime"`. This kind of matching usually works but is prone to
breakage if there are other gems that have the string `"sorbet-runtime"`
in their folder names, or if the `lib` folder is placed in a folder
that is named differently.

This commit improves the matching logic by storing the full path to the
`lib` folder in a constant, and then filtering caller location paths
that start with that path.

In order to get the full path to the `lib` folder, we need rely on the
const source location of the `T::Private` module, which is almost always
guaranteed to be defined in `lib/sorbet_runtime.rb`. The `File.dirname`
of that string gives us the full path to the `lib` folder for the
running Sorbet runtime gem.