Allow adding intrinsics to overloaded payload methods
Previously, Sorbet would only register the intrinsic on (effectively)
the placeholder method that we mangle rename out of the way to make way
for overloaded methods.

Also previously, we would not store intrinsics in the serialized global
state, presumably because it's a bad idea to store pointers there. But
now that we're just storing offsets into a fixed-size,
consistently-ordered vector, we don't have to do that and can just
serialize the intrinsic offset directly.