Add support for Ruby 3.2 Data Class
Ruby 3.2 introduced the [`Data`
class](https://bugs.ruby-lang.org/issues/16122), which is an immutable
data object similar to a `Struct`. This PR adds RBIs for the new `Data`
class as well as a rewriter to create correct types for the initializer
and add getter methods for the fields on a `Data` subclass.

Co-authored-by: Aiden Storey <aiden.storey@shopify.com>