Move private checking into dispatchCall
We were trying to detect the new/initialize mismatch using names only.
That proved to be too brittle in practice, specifically for the case
where the caller-side name is actually something like `<callWithBlock>`.
Moving the handling into dispatchCall makes it so that each intrinsic
can decide whether the recursive `dispatchCall` can have the call be
checked for visibility or not.