Encapsulate backtrace location searching logic to a single method
In order to provide compatibility with TruffleRuby that implements a lot
of its core library methods in Ruby, and for future versions of CRuby
that is increasingly doing the same, we need to be able to filter
all backtrace locations where the `path` starts with `<internal:`.

[This gist](https://gist.github.com/eregon/912e6359e83781c5fa1c638d3768c526)
shows the current state of the methods implemented in Ruby in CRuby,
JRuby and TruffleRuby.

Most recently [CRuby started implementing `Array#each` in Ruby](https://github.com/ruby/ruby/pull/9533), making it usages of `each` visible in backtraces with an `<internal:array>` path. This means that in order to be compatible with CRuby 3.4, Sorbet runtime needs to start filtering out backtrace locations that start with `<internal:`.

By encapsulating the caller location search logic into a singleton
method, we can apply that filtering in a single location and avoid
having to repeat it in multiple places.