Allow `T.attached_class` as argument of bounded generic (#6738)
* Allow `T.attached_class` as argument of bounded generic

Fix #6736

If you look what we do for other type members in type_syntax parsing, we
say that the `result.type` of the parse is a `SelfTypeParam` type. For
some reason, here we were using the `resultType` of the symbol itself,
which is a `LambdaParam`. Thus, this change does what we were doing for
the normal type member case, and that makes the test pass.

I don't have a great way to explain why we need a `SelfTypeParam` here
instead of a `LambdaParam`, because I've only learned what they mean the
same way a native speaker might learn grammar—the best I can do is say
something akin to it "sounds weird" trying to use `LambdaParam` here.

I can back up the weirdness in this particular case, because the code in
`isSubTypeUnderConstraint` doesn't actually implement the case for
`t1 <: t2` if only one of them is a `LambdaParam`. There are some
commented out ENFORCEs there that seem to suggest we wish it were the
case that we never even had code for `LambdaParam` there:

<https://github.com/sorbet/sorbet/blob/d4f8fa8d6f8639310d1af224324d3045fbc90dcc/core/types/subtyping.cc#L1135>

* `const TypeMemberRef &`