Only enter AttachedClass for classes, not modules
This means we don't have to allocate 8 MB worth of `TypeParameter`
objects on Stripe's codebase, and then another 3 MB worth of
`LambdaParam` stored inside those `TypeParameter`, up to another 2 MB
from not having to allocate and store `typeParams` for these modules
(upper bound, because maybe some of these modules have type_templates,
but that seems rare), and then one entry in the `members` of the
singleton class (which is an UnorderedMap, where I'm not really sure how
to think about space usage).

So conservatively, 10 MB - 13+ MB of memory saved, which is really a
drop in the bucket.